#5 cart php  http://pm.5psolutions.net/public/index.php?path_info=projects/training-andrei-stoleriu/tasks/5 checkout validations

+++ cart.php
@@ -24,7 +24,32 @@ if (!empty(array_values($_SESSION['cart']))) {
 $stmt->setFetchMode(PDO::FETCH_ASSOC);
 $rows = $stmt->fetchAll();
 
+
+
+$name = $contactDetails = $comments = '';
+$nameErr = $contactDetailsErr = $cartErr = '';
+
 if (isset($_POST['checkout'])) {
+
+    if (empty($_POST['name'])) {
+        $nameErr = __('Name is required');
+    } else {
+        $name = ($_POST['name']);
+    }
+    if (empty($_POST['contactDetails'])) {
+        $contactDetailsErr = __('Contact details are required');
+    } else {
+        $contactDetails = ($_POST['contactDetails']);
+    }
+
+    $comments = ($_POST['comments']);
+
+    if (empty($_SESSION['cart'])) {
+        $cartErr = __('Cart is empty');
+    }
+}
+
+if (isset($_POST['checkout']) && empty($nameErr) && empty($contactDetailsErr) && empty($cartErr)) {
 
     $name = $_POST['name'];
     $contactDetails = $_POST['contactDetails'];
@@ -68,13 +93,10 @@ if (isset($_POST['checkout'])) {
     if (mail($to, $subject, $message, $headers)) {
         header("refresh:5;url=cart.php");
         echo "<div class='p-3 mb-2 bg-primary text-white'>The email has been sent. Thank you" . " " . $name . "</div>";
-    } else {
-        header("refresh:5;url=cart.php");
-        echo "<div class='p-3 mb-2 bg-warning text-dark'>Something went wrong!</div>";
     }
 }
-?>
 
+?>
 <html lang="en">
 
 <head>
@@ -87,44 +109,47 @@ if (isset($_POST['checkout'])) {
 
 <body>
     <div class="container">
+        <?php if (empty($_SESSION['cart'])) : ?>
+            <h2 class="text-danger"> <?= $cartErr ?><h2>
+                <?php endif; ?>
+                <table class="table">
+                    <thead class="thead-dark">
+                        <tr>
+                            <th scope="col"></th>
+                            <th scope="col"><?= __('Title') ?></th>
+                            <th scope="col"><?= __('Description') ?></th>
+                            <th scope="col"><?= __('Price') ?></th>
+                            <th scope="col"><?= __('Action') ?></th>
+                        </tr>
+                    </thead>
+                    <?php foreach ($rows as $row) : ?>
+                        <tr>
+                            <td><img src="<?= $row['image'] ?>" style="width: 200px" alt=""></td>
+                            <td><?= $row['title'] ?></td>
+                            <td><?= $row['description'] ?></td>
+                            <td> $ <?= $row['price'] ?></td>
+                            <td><a href="?delete=<?= $row['id'] ?>"><?= __('Delete') ?></a></td>
+                        </tr>
+                    <?php endforeach; ?>
 
-        <table class="table">
-            <thead class="thead-dark">
-                <tr>
-                    <th scope="col"></th>
-                    <th scope="col"><?= __('Title') ?></th>
-                    <th scope="col"><?= __('Description') ?></th>
-                    <th scope="col"><?= __('Price') ?></th>
-                    <th scope="col"><?= __('Action') ?></th>
-                </tr>
-            </thead>
-            <?php foreach ($rows as $row) : ?>
-                <tr>
-                    <td><img src="<?= $row['image'] ?>" style="width: 200px" alt=""></td>
-                    <td><?= $row['title'] ?></td>
-                    <td><?= $row['description'] ?></td>
-                    <td> $ <?= $row['price'] ?></td>
-                    <td><a href="?delete=<?= $row['id'] ?>"><?= __('Delete') ?></a></td>
-                </tr>
-            <?php endforeach; ?>
+                </table>
+                <form class="form-group" method="POST" action="cart.php">
+                    <label for="name"><?= __('Name') ?></label>
+                    <input type="text" name="name" value="" placeholder="<?= __('Insert your name') ?>" class="form-control" value="<?php echo $name ?>">
+                    <p class="text-danger"> <?= $nameErr; ?></p>
+                    <label for="contactDetails"><?= __('Contact details') ?></label>
+                    <textarea rows="2" cols="30" name="contactDetails" value="" placeholder="<?= __('Insert your contact details') ?>" class="form-control" value="<?php echo $contactDetails ?>"></textarea>
+                    <p class="text-danger"> <?= $contactDetailsErr ?></p>
+                    <label for="comments"><?= __('Comments') ?></label>
+                    <textarea rows="4" cols="30" name="comments" value="" placeholder="<?= __('Insert comments') ?>" class="form-control" value="<?php echo $comments ?>"></textarea>
+                    <input type="submit" class="btn btn-primary" name="checkout" value="<?= __('Checkout') ?>"></button>
+                </form>
 
-        </table>
+                <a href="index.php"><?= __('Go to index') ?></a>
 
-        <form class="form-group" method="POST" action="cart.php">
-            <label for="name"><?= __('Name') ?></label>
-            <input type="text" name="name" value="" placeholder="<?= __('Insert your name') ?>" class="form-control" value="<?php echo $name ?>"> <br />
-            <label for="contactDetails"><?= __('Contact details') ?></label>
-            <textarea rows="2" cols="30" name="contactDetails" value="" placeholder="<?= __('Insert your contact details') ?>" class="form-control" value="<?php echo $contactDetails ?>"></textarea> <br />
-            <label for="comments"><?= __('Comments') ?></label>
-            <textarea rows="4" cols="30" name="comments" value="" placeholder="<?= __('Insert comments') ?>" class="form-control" value="<?php echo $comments ?>"></textarea> <br />
-            <input type="submit" class="btn btn-primary" name="checkout" value="<?= __('Checkout') ?>"></button>
-        </form>
-
-        <a href="index.php"><?= __('Go to index') ?></a>
-
-        <script src="https://code.jquery.com/jquery-3.4.1.slim.min.js" integrity="sha384-J6qa4849blE2+poT4WnyKhv5vZF5SrPo0iEjwBvKU7imGFAV0wwj1yYfoRSJoZ+n" crossorigin="anonymous"></script>
-        <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js" integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo" crossorigin="anonymous"></script>
-        <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js" integrity="sha384-wfSDF2E50Y2D1uUdj0O3uMBJnjuUD4Ih7YwaYd1iqfktj0Uod8GCExl3Og8ifwB6" crossorigin="anonymous"></script>
+                <script src="https://code.jquery.com/jquery-3.4.1.slim.min.js" integrity="sha384-J6qa4849blE2+poT4WnyKhv5vZF5SrPo0iEjwBvKU7imGFAV0wwj1yYfoRSJoZ+n" crossorigin="anonymous"></script>
+                <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js" integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo" crossorigin="anonymous"></script>
+                <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js" integrity="sha384-wfSDF2E50Y2D1uUdj0O3uMBJnjuUD4Ih7YwaYd1iqfktj0Uod8GCExl3Og8ifwB6" crossorigin="anonymous"></script>
     </div>
 </body>
